% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadEro_run.R
\name{RadEro_run}
\alias{RadEro_run}
\title{Starts conversion model for soil redistribution}
\usage{
RadEro_run(data, config, AxisMaxValue = NULL)
}
\arguments{
\item{data}{A data frame containing the Cs137 inventory data from soil profiles. It must be located in the working directory.}

\item{config}{A list or configuration file that defines the parameters for the model run. It must be located in the working directory.}

\item{AxisMaxValue}{Optional. A numeric value specifying the maximum value for the axes in the plot. Defaults to NULL.}
}
\value{
Each soil profile analysed will generate a  plot.png and a temp folder with its corresponding ID. In addition, the whole set of results of all the points will be reported in a results.txt summary file in the working directory.
}
\description{
Initiates the model based on the Cs137 inventory data from soil profiles according to the specified configuration in the working directory.
}
