% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadEro_run.R
\name{RadEro_run}
\alias{RadEro_run}
\title{Starts conversion model}
\usage{
RadEro_run(data, config, AxisMaxValue = NULL)
}
\description{
Initiates the model acording to the inventory data of soil profiles according to the specified configuration.
}
